+++ src/integrations/anthropic_api.py
@@ -22,9 +22,9 @@ from anthropic.types import Message, ModelParam, TextBlock, ToolUseBlock
 from tenacity import after_log, before_log, retry, retry_if_exception, stop_after_attempt, wait_exponential
 
 # Local Application Imports
-from src.config import Config, logger
-from src.constants import CLIENT_ERROR_CODE, GENERIC_API_ERROR_MESSAGE, SERVER_ERROR_CODE
-from src.utils import truncate_text, validate_env_var
+from src.common.config import Config, logger
+from src.common.constants import CLIENT_ERROR_CODE, GENERIC_API_ERROR_MESSAGE, SERVER_ERROR_CODE
+from src.common.utils import truncate_text, validate_env_var
 
 PROMPT_TEMPLATE: str = """
 <role>
