+++ src/integrations/elevenlabs_api.py
@@ -25,9 +25,9 @@ from elevenlabs.core import ApiError
 from tenacity import after_log, before_log, retry, retry_if_exception, stop_after_attempt, wait_fixed
 
 # Local Application Imports
-from src.config import Config, logger
-from src.constants import CLIENT_ERROR_CODE, GENERIC_API_ERROR_MESSAGE, SERVER_ERROR_CODE
-from src.utils import save_base64_audio_to_file, validate_env_var
+from src.common.config import Config, logger
+from src.common.constants import CLIENT_ERROR_CODE, GENERIC_API_ERROR_MESSAGE, SERVER_ERROR_CODE
+from src.common.utils import save_base64_audio_to_file, validate_env_var
 
 
 @dataclass(frozen=True)
