+++ src/integrations/hume_api.py
@@ -24,9 +24,9 @@ from hume.tts.types import Format, FormatMp3, PostedUtterance, ReturnTts
 from tenacity import after_log, before_log, retry, retry_if_exception, stop_after_attempt, wait_fixed
 
 # Local Application Imports
-from src.config import Config, logger
-from src.constants import CLIENT_ERROR_CODE, GENERIC_API_ERROR_MESSAGE, RATE_LIMIT_ERROR_CODE, SERVER_ERROR_CODE
-from src.utils import save_base64_audio_to_file, validate_env_var
+from src.common.config import Config, logger
+from src.common.constants import CLIENT_ERROR_CODE, GENERIC_API_ERROR_MESSAGE, RATE_LIMIT_ERROR_CODE, SERVER_ERROR_CODE
+from src.common.utils import save_base64_audio_to_file, validate_env_var
 
 
 @dataclass(frozen=True)
