+++ src/integrations/openai_api.py
@@ -25,9 +25,9 @@ from openai import APIError, AsyncOpenAI
 from tenacity import after_log, before_log, retry, retry_if_exception, stop_after_attempt, wait_fixed
 
 # Local Application Imports
-from src.config import Config, logger
-from src.constants import CLIENT_ERROR_CODE, GENERIC_API_ERROR_MESSAGE, SERVER_ERROR_CODE
-from src.utils import validate_env_var
+from src.common.config import Config, logger
+from src.common.constants import CLIENT_ERROR_CODE, GENERIC_API_ERROR_MESSAGE, SERVER_ERROR_CODE
+from src.common.utils import validate_env_var
 
 
 @dataclass(frozen=True)
