+++ src/middleware/__init__.py
@@ -1,0 +1,3 @@
+from src.middleware.meta_tag_injection import MetaTagInjectionMiddleware
+
+__all__ = ["MetaTagInjectionMiddleware"]
